backend/requirements.txt
# ===========================================
# Backend Dependencies
# ===========================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
slowapi==0.1.9

# Request Handling
python-multipart==0.0.6

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Validation
email-validator==2.1.0

# Caching & Task Queue
redis==5.0.1
celery==5.3.4
APScheduler==3.10.4

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Report Generation
qrcode[pil]==7.4.2
Pillow==10.1.0
reportlab==4.0.7
openpyxl==3.1.2

# HTTP Client (for health checks)
httpx==0.25.2

# Testing (dev only)
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Gunicorn (production WSGI server)
gunicorn==21.2.0
